Add first version of acceptance test. Modify main to pass the test.
The implementation is far from complete. The acceptance test
'intercepts' standard output from App.main and verifies that the team
with the least goal difference is printed to the console.

The team is Aston Villa, according to the test file 'football.csv'.
App.main is modified to print 'Aston_Villa' in order to test the
acceptance test.