Start to fix acceptance test.
The first acceptance test expected a single football team as having the
lowest absolute goal difference. In the test file, Aston Villa is the
single team with the lowest goal difference. It is possible that more
than one team share the lowest goal difference.

An unmodified copy of the original football.csv was added to the test
resources. Its name is football-without-dups.csv. The acceptance test
now uses this copy.

A modified version of the file was with the name football-with-dups.csv.
The goals and goals allowed were set so that West Ham, Aston Villa and
Southampton share the same absolute goal difference with West Ham having
a positive difference and the other a negative difference. A second
acceptance test was added for this case.

Needless to say, the new acceptance test fails due to this same
oversight in the implementation.

Next step: refactor the implementation.