Add class for keeping objects with common minimum value.
This change is a step in making the code more portable. The class
generalizes the common steps in keeping a list of objects with a common
minimum value. It can be re-usued for any data type on any Comparable
field or handle any arbitrary ranking function.

It also a step in separating the concerns of the classes in the
application.