Add data structure for football standings.
The intention is to use this class instead of a raw row from a CSV file.
IMO the code is a little easier to follow with a football standings
class than it is with an array of fields.

The object only contains the fields that are relevant to the task. The
other fields in the test files could be added if there is enough time.